pyproject.toml
[tool.poetry]
name = "slap"
version = "0.1.0"
description = "SlapMac clone in Python"
authors = ["elevitpachon"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
sounddevice = "^0.5.0"
numpy = "^2.0"
rumps = {version = "^0.4.0", markers = "sys_platform == 'darwin'"}
pystray = {version = "^0.19", markers = "sys_platform == 'linux'"}
Pillow = {version = "^11.0", markers = "sys_platform == 'linux'"}
pygame = {version = "^2.6", markers = "sys_platform == 'linux'"}

[tool.poetry.group.dev.dependencies]
ipython = "^8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
